Backend

+++ src/main/java/hello/inven/helloinven/serviceimpl/ItemServiceImpl.java
@@ -1,5 +1,6 @@
 package hello.inven.helloinven.serviceimpl;
 
+import hello.inven.helloinven.exceptionhandler.BadRequestException;
 import hello.inven.helloinven.exceptionhandler.NotFoundException;
 import hello.inven.helloinven.model.Item;
 import hello.inven.helloinven.response.ResponseAjax;
@@ -28,50 +29,42 @@ public class ItemServiceImpl implements ItemService {
     CategoryRepository categoryRepository;
 
     @Override
-    public ResponseAjax createItem(Item item) throws IOException {
-//    public ResponseAjax createItem(Item item, MultipartFile file) throws IOException {
-        Item newItem = new Item();
-        System.out.println("item ID: " + item.getId() + "item Name: " + item.getName());
-        newItem.setId(item.getId());
-        newItem.setName(item.getName());
-        newItem.setItemType(Item.ItemType.ITEM);
-        newItem.setQuantity(item.getQuantity());
-        newItem.setPrice(item.getPrice());
-        newItem.setWeight(item.getWeight());
-        newItem.setHeight(item.getHeight());
-        newItem.setWidth(item.getWidth());
-        newItem.setDepth(item.getDepth());
+    public Item createItem(Item item) throws IOException {
+        Item existingItem = itemRepository.findById(item.getId()).orElse(null);
+        if (existingItem == null) {
+            Item newItem = new Item();
+            newItem.setId(item.getId());
+            newItem.setName(item.getName());
+            newItem.setItemType(Item.ItemType.ITEM);
+            newItem.setQuantity(item.getQuantity());
+            newItem.setPrice(item.getPrice());
+            newItem.setWeight(item.getWeight());
+            newItem.setHeight(item.getHeight());
+            newItem.setWidth(item.getWidth());
+            newItem.setDepth(item.getDepth());
 
-//        Integer categoryId = item.getCategory().getId();
-        System.out.println("Category ID apakah dapat: " + item.getCategory().getId());
-        System.out.println("Ini cuma category: "+ item.getCategory());
-//        Category category = categoryRepository.findById(categoryId).get();
+            newItem.setCategory(item.getCategory());
 
+            MultipartFile file = item.getImage();
 
-//        newItem.setCategory(category);
-        newItem.setCategory(item.getCategory());
+            if (!file.isEmpty()) {
+                String fileName = file.getOriginalFilename();
+                InputStream is = file.getInputStream();
+                String uploadDirectory = System.getProperty("user.dir") + "/uploads/item/";
 
-        MultipartFile file = item.getImage();
+                try {
+                    Files.copy(is, Paths.get(uploadDirectory + fileName).toAbsolutePath().normalize(), StandardCopyOption.REPLACE_EXISTING);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
 
-        if(!file.isEmpty()){
-            String fileName = file.getOriginalFilename();
-            InputStream is = file.getInputStream();
-//            String uploadDirectory = System.getProperty("user.dir") + "/src/main/resources/static/uploads/item/";
-            String uploadDirectory = System.getProperty("user.dir") + "/uploads/item/";
-
-            try {
-                Files.copy(is, Paths.get(uploadDirectory + fileName).toAbsolutePath().normalize(), StandardCopyOption.REPLACE_EXISTING);
-            }
-            catch (IOException e){
-                e.printStackTrace();
+                newItem.setImagePath(fileName);
             }
 
-            newItem.setImagePath(fileName);
-            System.out.println("\nFileSUDAHDITAMBAH: " + Paths.get(uploadDirectory + fileName).toAbsolutePath().normalize());
+            itemRepository.save(newItem);
+            return item;
         }
-
-         itemRepository.save(newItem);
-        return new ResponseAjax("Created", newItem);
+        else throw new BadRequestException("Item ID is already exists!");
     }
 
     @Override
@@ -81,11 +74,10 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public ResponseAjax deleteItem(Long id){
+    public Item deleteItem(Long id){
         Item item = itemRepository.findById(id).orElse(null);
-        System.out.println(item);
         if (item != null) {
-            if (!item.getImagePath().isEmpty()){
+            if (item.getImagePath() != null){
                 String photoDirectory = System.getProperty("user.dir") + "/uploads/item/";
                 try {
                     File file = new File(photoDirectory + item.getImagePath());
@@ -101,10 +93,9 @@ public class ItemServiceImpl implements ItemService {
             }
 
             itemRepository.delete(item);
-            return new ResponseAjax("Deleted", "Item has been deleted!");
+            return item;
         }
         else {
-//            return new ResponseAjax("Failed", "Item is failed to be deleted!");
             throw new NotFoundException("Item not found and failed to be deleted!");
         }
 
@@ -112,7 +103,6 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Item detailItem(Long id){
-//            public Item detailItem(Long id){
         Item item = itemRepository.findById(id).orElse(null);
         if (item != null){
             return item;
